+++ src/main/java/com/projectsassy/sassy/chatting/controller/ChatController.java
@@ -1,21 +1,17 @@
-package com.projectsassy.sassy.websoket;
+package com.projectsassy.sassy.chatting.controller;
 
-import com.projectsassy.sassy.chattingroom.domain.ChattingRoom;
-import com.projectsassy.sassy.common.code.ErrorCode;
-import com.projectsassy.sassy.common.exception.UnauthorizedException;
-import com.projectsassy.sassy.user.domain.SessionConst;
+import com.projectsassy.sassy.chatting.service.ChattingRoomService;
+import com.projectsassy.sassy.chatting.service.MessageService;
 import com.projectsassy.sassy.user.domain.User;
 import com.projectsassy.sassy.user.service.UserService;
-import com.projectsassy.sassy.websoket.second.MatchRequest;
-import com.projectsassy.sassy.websoket.second.MbtiCondition;
-import com.projectsassy.sassy.websoket.second.MessageRequest2;
-import com.projectsassy.sassy.websoket.second.WaitingRequest;
+import com.projectsassy.sassy.chatting.dto.MatchRequest;
+import com.projectsassy.sassy.chatting.data.MatchingMbtiData;
+import com.projectsassy.sassy.chatting.dto.MessageRequest2;
+import com.projectsassy.sassy.chatting.dto.WaitingRequest;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.messaging.handler.annotation.DestinationVariable;
-import org.springframework.messaging.handler.annotation.Header;
 import org.springframework.messaging.handler.annotation.MessageMapping;
-import org.springframework.messaging.simp.SimpMessageHeaderAccessor;
 import org.springframework.messaging.simp.SimpMessageSendingOperations;
 import org.springframework.web.bind.annotation.*;
 
@@ -23,6 +19,7 @@ import java.util.concurrent.ConcurrentHashMap;
 
 @RestController
 @RequiredArgsConstructor
+@Slf4j
 public class ChatController {
 
     private final SimpMessageSendingOperations simpMessageSendingOperations;
@@ -30,42 +27,47 @@ public class ChatController {
     private final MessageService messageService;
     private final UserService userService;
 
-    @PostMapping("/chat")
-    public ResponseEntity createChattingRoom(
-        @RequestBody ChatRequest chatRequest,
-        @SessionAttribute(name = SessionConst.USER_ID, required = false) Long loginUserId
-    ) {
-        validateUser(loginUserId);
-        ChattingRoom chattingRoom = chattingRoomService.findRoomOrCreate(chatRequest, loginUserId);
+//    @PostMapping("/chat")
+//    public ResponseEntity createChattingRoom(
+//        @RequestBody ChatRequest chatRequest,
+//        @SessionAttribute(name = SessionConst.USER_ID, required = false) Long loginUserId
+//    ) {
+//        validateUser(loginUserId);
+//        ChattingRoom chattingRoom = chattingRoomService.findRoomOrCreate(chatRequest, loginUserId);
+//
+//        return ResponseEntity.ok().body(chattingRoom.getId());
+//    }
+//
+//    @MessageMapping("/chat/{roomId}")
+//    public void message(@DestinationVariable("roomId") Long roomId, MessageRequest messageRequest) {
+//        //messageRequest 는 어떤값을 넣을지 정해야할듯.
+//        String type = messageRequest.getType();
+//        if (type.equals("close")) {
+//            chattingRoomService.deleteEmptyRoom(messageRequest);
+//            simpMessageSendingOperations.convertAndSend("/sub/chat/" + roomId, messageRequest);
+//            return;
+//        }
+//
+//        simpMessageSendingOperations.convertAndSend("/sub/chat/" + roomId, messageRequest);
+//
+//        messageService.saveMessage(roomId, messageRequest);
+//
+//    }
 
-        return ResponseEntity.ok().body(chattingRoom.getId());
-    }
+//    private static void validateUser(Long loginUserId) {
+//        if (loginUserId == null) {
+//            throw new UnauthorizedException(ErrorCode.UNAUTHORIZED);
+//        }
+//    }
 
-    @MessageMapping("/chat/{roomId}")
-    public void message(@DestinationVariable("roomId") Long roomId, MessageRequest messageRequest) {
-        //messageRequest 는 어떤값을 넣을지 정해야할듯.
-        String type = messageRequest.getType();
-        if (type.equals("close")) {
-            chattingRoomService.deleteEmptyRoom(messageRequest);
-            simpMessageSendingOperations.convertAndSend("/sub/chat/" + roomId, messageRequest);
-            return;
-        }
-
-        simpMessageSendingOperations.convertAndSend("/sub/chat/" + roomId, messageRequest);
-
-        messageService.saveMessage(roomId, messageRequest);
-
-    }
-
-    private static void validateUser(Long loginUserId) {
-        if (loginUserId == null) {
-            throw new UnauthorizedException(ErrorCode.UNAUTHORIZED);
-        }
-    }
-
-    private ConcurrentHashMap<Long, MbtiCondition> waiting = new ConcurrentHashMap<>();
+    private ConcurrentHashMap<Long, MatchingMbtiData> waiting = new ConcurrentHashMap<>();
 
     //2번째 방법.
+    // ws://localhost:8080/ws/pub/chat/wait  WaitingRequest : type:일단 wait, close 두개로. userId, selectMbti
+    // ws://localhost:8080/ws/sub/chat/wait/{userId}  : MatchRequest : type:match, roomId;
+    // 둘이 매칭 돼서 roomId를 받음.
+    // ws://localhost:8080/ws/pub/chat/match/{roomId}  MessageRequest2 : type:채팅(대화), 나갔을때경우 , roomId, sendUserId, content
+    // ws://localhost:8080/ws/sub/chat/match/{roomId}  MessageRequest2 : type:채팅(대화), 나갔을때경우 , roomId, sendUserId, content
     @MessageMapping("/chat/wait")  // 유저는 자신의 userId인 /sub/chat/wait/{userId} 에서 대기.
     public void addUserToWaiting(WaitingRequest waitingRequest) { // type:일단 wait, close 두개로. userId, selectMbti
         Long userId = Long.valueOf(waitingRequest.getUserId());
@@ -82,8 +84,9 @@ public class ChatController {
         String selectMbti = waitingRequest.getSelectMbti();
 
         //나의 mbti와 내가 원하는 mbti로 대기 줄 waiting에 mbti 조건으로 추가. -> 이 조건을 이용하여 매칭.
-        MbtiCondition mbtiCondition = new MbtiCondition(myMbti, selectMbti);
-        waiting.put(userId, mbtiCondition);
+        MatchingMbtiData matchingMbtiData = new MatchingMbtiData(myMbti, selectMbti);
+        waiting.put(userId, matchingMbtiData);
+        log.info("size={}", waiting.size());
 
 //        로그
 //        System.out.println(waiting.size());
@@ -97,9 +100,9 @@ public class ChatController {
                 //waiting map에서 for 문을 이용해 탐색할 때 탐색하는 값이 '나' 라면 스킵.
                 if (userId == waitingUserId) continue;
 
-                MbtiCondition waitingMbtiCondition = waiting.get(waitingUserId);
-                String waitingUserMbti = waitingMbtiCondition.getMyMbti();
-                String waitingUserSelectMbti = waitingMbtiCondition.getSelectMbti();
+                MatchingMbtiData waitingMatchingMbtiData = waiting.get(waitingUserId);
+                String waitingUserMbti = waitingMatchingMbtiData.getMyMbti();
+                String waitingUserSelectMbti = waitingMatchingMbtiData.getSelectMbti();
 
                 if (selectMbti.equals(waitingUserMbti) && myMbti.equals(waitingUserSelectMbti)) {
                     Long roomId = chattingRoomService.createChattingRoom(user, waitingUserId);
@@ -111,6 +114,7 @@ public class ChatController {
                     simpMessageSendingOperations.convertAndSend("/sub/chat/wait/" + waitingUserId, matchRequest);
                     waiting.remove(userId);
                     waiting.remove(waitingUserId);
+                    log.info("size={}", waiting.size());
                     break;
                 }
