+++ src/main/java/com/projectsassy/sassy/chatting/dto/MatchRequest.java
@@ -1,15 +1,16 @@
-package com.projectsassy.sassy.websoket.second;
+package com.projectsassy.sassy.chatting.dto;
 
 import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Getter
+@AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class WaitingRequest {
+public class MatchRequest {
 
     private String type;
-    private String userId;
-    private String selectMbti;
+    private String roomId;
 
 }
