+++ src/main/java/com/projectsassy/sassy/chatting/dto/MessageRequest2.java
@@ -1,18 +1,16 @@
-package com.projectsassy.sassy.websoket;
+package com.projectsassy.sassy.chatting.dto;
 
 import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Getter
-@AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class MessageRequest {
+public class MessageRequest2 {
 
     private String type;
-    private String sendUserId;
     private String roomId;
+    private String sendUserId;
     private String content;
 
 }
