+++ src/main/java/com/projectsassy/sassy/chatting/dto/WaitingRequest.java
@@ -1,4 +1,4 @@
-package com.projectsassy.sassy.websoket.second;
+package com.projectsassy.sassy.chatting.dto;
 
 import lombok.AccessLevel;
 import lombok.Getter;
@@ -6,11 +6,10 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class MessageRequest2 {
+public class WaitingRequest {
 
     private String type;
-    private String roomId;
-    private String sendUserId;
-    private String content;
+    private String userId;
+    private String selectMbti;
 
 }
