+++ src/main/java/com/projectsassy/sassy/chatting/repository/ChattingRoomRepository.java
@@ -1,6 +1,6 @@
-package com.projectsassy.sassy.chattingroom.repository;
+package com.projectsassy.sassy.chatting.repository;
 
-import com.projectsassy.sassy.chattingroom.domain.ChattingRoom;
+import com.projectsassy.sassy.chatting.domain.ChattingRoom;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
